Reorganise identity-backfill-batch by symbol type (interfaces, types, schemas, functions)
