+++ modules/identity-backfill-batch/src/functions/csvRow.ts
@@ -1,0 +1,5 @@
+import { escapeCsv } from './escapeCsv';
+
+export function csvRow(values: string[]): string {
+	return values.map(escapeCsv).join(',') + '\n';
+}
