+++ modules/identity-backfill-batch/src/functions/ensureHeader.ts
@@ -1,0 +1,7 @@
+import { existsSync, writeFileSync } from 'node:fs';
+
+export function ensureHeader(path: string, header: string): void {
+	if (!existsSync(path)) {
+		writeFileSync(path, header);
+	}
+}
