+++ modules/identity-backfill-batch/src/functions/escapeCsv.ts
@@ -1,0 +1,6 @@
+export function escapeCsv(v: string): string {
+	if (v.includes(',') || v.includes('"') || v.includes('\n')) {
+		return `"${v.replace(/"/g, '""')}"`;
+	}
+	return v;
+}
