+++ modules/identity-backfill-batch/src/functions/formatDuration.ts
@@ -1,0 +1,5 @@
+export function formatDuration(ms: number): string {
+	const s = Math.floor(ms / 1000);
+	const m = Math.floor(s / 60);
+	return `${m}m ${s % 60}s`;
+}
