+++ modules/identity-backfill-batch/src/functions/loadState.ts
@@ -1,0 +1,17 @@
+import { existsSync, readFileSync } from 'node:fs';
+import { join } from 'node:path';
+import type { IState } from '../interfaces';
+import { stateFileSchema } from '../schemas';
+
+export function loadState(outputDir: string): IState {
+	const path = join(outputDir, 'state.json');
+	if (!existsSync(path)) {
+		return { processed: new Set(), rejected: new Set(), errored: new Set() };
+	}
+	const raw = stateFileSchema.parse(JSON.parse(readFileSync(path, 'utf8')));
+	return {
+		processed: new Set(raw.processed ?? []),
+		rejected: new Set(raw.rejected ?? []),
+		errored: new Set(raw.errored ?? []),
+	};
+}
