+++ modules/identity-backfill-batch/src/functions/pickEmail.ts
@@ -1,0 +1,5 @@
+import type { ICsvRow } from '../interfaces';
+
+export function pickEmail(row: ICsvRow): string | null {
+	return row.sf_contact_email || row.zuora_bill_to_email || null;
+}
