+++ modules/identity-backfill-batch/src/interfaces/IApiConfig.ts
@@ -1,0 +1,4 @@
+export interface IApiConfig {
+	url: string;
+	apiKey: string;
+}
