+++ modules/identity-backfill-batch/src/interfaces/ICsvRow.ts
@@ -1,0 +1,11 @@
+export interface ICsvRow {
+	sub_id: string;
+	sub_number: string;
+	subscription_end_date: string;
+	sub_status: string;
+	zuora_bill_to_email: string;
+	sf_contact_email: string;
+	identity_status: string;
+	sf_product: string;
+	account_crm_id: string;
+}
