+++ modules/identity-backfill-batch/src/interfaces/IState.ts
@@ -1,0 +1,5 @@
+export interface IState {
+	processed: Set<string>;
+	rejected: Set<string>;
+	errored: Set<string>;
+}
