+++ modules/identity-backfill-batch/src/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from './IApiConfig';
+export * from './IArgs';
+export * from './ICsvRow';
+export * from './IState';
