+++ modules/identity-backfill-batch/src/schemas/filterEnum.ts
@@ -1,0 +1,3 @@
+import { z } from 'zod';
+
+export const filterEnum = z.enum(['has-identity-id', 'no-identity-id', 'all']);
