+++ modules/identity-backfill-batch/src/schemas/responseBodySchema.ts
@@ -1,0 +1,3 @@
+import { z } from 'zod';
+
+export const responseBodySchema = z.object({ message: z.string().optional() });
