+++ modules/identity-backfill-batch/src/types/ApiOutcome.ts
@@ -1,0 +1,4 @@
+export type ApiOutcome =
+	| { kind: 'success'; identityId: string | null; rawBody: string }
+	| { kind: 'rejected'; reason: string; httpStatus: number }
+	| { kind: 'error'; reason: string; httpStatus: number | null };
