+++ modules/identity-backfill-batch/src/types/FilterMode.ts
@@ -1,0 +1,1 @@
+export type FilterMode = 'has-identity-id' | 'no-identity-id' | 'all';
