+++ modules/identity-backfill-batch/src/types/Stage.ts
@@ -1,0 +1,1 @@
+export type Stage = 'CODE' | 'PROD';
