+++ modules/identity-backfill-batch/src/types/WriteMode.ts
@@ -1,0 +1,1 @@
+export type WriteMode = 'dry-run-only' | 'real';
