+++ modules/identity-backfill-batch/src/types/index.ts
@@ -1,0 +1,4 @@
+export * from './ApiOutcome';
+export * from './FilterMode';
+export * from './Stage';
+export * from './WriteMode';
